requirements.txt
tensorflow>=2.11
torch
numpy
pillow
opencv-python
matplotlib
tqdm
